Updating build copy

+++ marlin-2.0.9.1/skr-14-turbo/build_copy.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
+if [ ! -d "/media/alex/disk" ]
+then
+    echo "SD mount point doesn't exist"
+    exit 1
+fi
 rm /media/alex/disk/*
 pio run
 cp .pio/build/LPC1769/firmware.bin /media/alex/disk/
